Refs #145. Fixed issues observed during testing picking single peaks.
