+++ pyvdrive/interface/PeakPickWindow.py
@@ -1,383 +1,34 @@
 ########################################################################
 #
-# Window for set up log slicing splitters
+# Window for processing peaks
 #
 ########################################################################
-import sys
 import os
 try:
     import qtconsole.inprocess
     from PyQt5 import QtCore as QtCore
     from PyQt5.QtWidgets import QVBoxLayout
     from PyQt5.uic import loadUi as load_ui
-    from PyQt5.QtWidgets import QDialog, QMainWindow, QLineEdit, QComboBox, QCheckBox, QMainWindow, QButtonGroup
+    from PyQt5.QtWidgets import QMainWindow, QButtonGroup
     from PyQt5.QtWidgets import QFileDialog
 except ImportError:
     from PyQt4 import QtCore as QtCore
     from PyQt4.QtGui import QVBoxLayout
     from PyQt4.uic import loadUi as load_ui
-    from PyQt4.QtGui import QDialog, QMainWindow, QLineEdit, QComboBox, QCheckBox, QMainWindow, QButtonGroup
+    from PyQt4.QtGui import QMainWindow, QButtonGroup
     from PyQt4.QtGui import QFileDialog
-
 try:
     _fromUtf8 = QtCore.QString.fromUtf8
 except AttributeError:
     def _fromUtf8(s):
         return s
-
 import gui.GuiUtility as GuiUtility
-# from pyvdrive.interface.gui.diffractionplotview import DiffractionPlotView
-# from pyvdrive.interface.gui.vdrivetablewidgets import PeakParameterTable
-# from pyvdrive.interface.gui.vdrivetreewidgets import SinglePeakFitManageTree
 import gui.diffractionplotview as dv
-import GroupPeakDialog
+from peak_processing_helper import GroupPeakDialog, PeakWidthSetupDialog, PhaseWidgets, UnitCellList
 from pyvdrive.interface.gui.diffractionplotview import DiffractionPlotView
 from pyvdrive.interface.gui.vdrivetablewidgets import PeakParameterTable
-from pyvdrive.interface.gui.vdrivetreewidgets import SinglePeakFitManageTree
 import vanadium_controller_dialog
-
-# Set up path to PyVDrive
-import socket
-# if it is on analysis computer...
-# if socket.gethostname().count('analysis-') > 0 or os.path.exists('/home/wzz') is False:
-#     sys.path.append('/SNS/users/wzz/local/lib/python/site-packages/')
-
-
 import pyvdrive.lib.peak_util as peak_util
-
-
-# List of supported unit cell
-UnitCellList = [('BCC', 'I m -3 m'),
-                ('FCC', 'F d -3 m'),
-                ('HCP', 'P 63/m m c'),
-                ('Body-Center', 'I m m m'),
-                ('Face-Center', 'F m m m'),
-                ('Primitive', 'P m m m')]
-
-
-class PeakWidthSetupDialog(QDialog):
-    """
-    Class for set up dialog
-    """
-    def __init__(self, parent):
-        """
-        Init ...
-        :return:
-        """
-
-        # Initialize
-        QDialog.__init__(self, parent)
-
-        self.ui = load_ui("gui/PeakWidthSetup.ui", baseinstance=self)
-        self = self._promote_widgets()
-
-        # Define event handlers
-        self.ui.pushButton_cancel.clicked.connect(self.do_quit)
-        self.ui.pushButton_set.clicked.connect(self.do_set_width)
-
-        # self.connect(self.ui.pushButton_cancel, QtCore.SIGNAL('clicked()'),
-        #              self.do_quit)
-        # self.connect(self.ui.pushButton_set, QtCore.SIGNAL('clicked()'),
-        #              self.do_set_width)
-
-        # Class variables
-        self._peakWidth = None
-
-        return
-
-    def _promote_widgets(self):
-        treeView_iptsRun_layout = QVBoxLayout()
-        self.ui.frame_treeView_iptsRun.setLayout(treeView_iptsRun_layout)
-        self.ui.treeView_iptsRun = SinglePeakFitManageTree(self)
-        treeView_iptsRun_layout.addWidget(self.ui.treeView_iptsRun)
-
-        graphicsView_main_layout = QVBoxLayout()
-        self.ui.frame_graphicsView_main.setLayout(graphicsView_main_layout)
-        self.ui.graphicsView_main = DiffractionPlotView(self)
-        graphicsView_main_layout.addWidget(self.ui.graphicsView_main)
-
-        tableWidget_peakParameter_layout = QVBoxLayout()
-        self.ui.frame_tableWidget_peakParameter.setLayout(tableWidget_peakParameter_layout)
-        self.ui.tableWidget_peakParameter = PeakParameterTable(self)
-        tableWidget_peakParameter_layout.addWidget(self.ui.tableWidget_peakParameter)
-
-        return
-
-    def do_quit(self):
-        """
-        Return without doing anything
-        :return:
-        """
-        self.close()
-
-    def do_set_width(self):
-        """
-        Set peak width
-        :return:
-        """
-        peak_width = GuiUtility.parse_float(self.ui.lineEdit_peakWidth)
-        if peak_width is None:
-            GuiUtility.pop_dialog_error(self, 'Peak width is not set up!')
-            return
-        if peak_width <= 0.:
-            GuiUtility.pop_dialog_error(self, 'Peak width %f cannot be 0 or negative!' % peak_width)
-            return
-
-        self._peakWidth = peak_width
-
-        # Close
-        self.close()
-
-        return
-
-    def get_peak_width(self):
-        """ Get peak width
-        Purpose: Get the stored peak width from the window object
-        Requirements: it must be set up
-        Guarantees: the peak width is given if it is set up
-        :return:
-        """
-        assert self._peakWidth
-
-        return self._peakWidth
-
-
-class PhaseWidgets(object):
-    """
-    A set of widgets to define a phase
-    """
-    def __init__(self, parent, edit_a, edit_b, edit_c, edit_name, combo_box_type, check_box_select):
-        """
-        Initialize the phase widgets group
-        Requirements: all the inputs should be the proper PyQt widgets
-        :return:
-        """
-        # Check requirements
-        assert isinstance(parent, QMainWindow)
-        assert isinstance(edit_a, QLineEdit)
-        assert isinstance(edit_b, QLineEdit)
-        assert isinstance(edit_c, QLineEdit)
-        assert isinstance(edit_name, QLineEdit)
-        assert isinstance(combo_box_type, QComboBox)
-        assert isinstance(check_box_select, QCheckBox)
-
-        # Lattice parameters
-        self._lineEdit_a = edit_a
-        self._lineEdit_b = edit_b
-        self._lineEdit_c = edit_c
-        # Phase' name
-        self._lineEdit_name = edit_name
-        # Phase' type
-        self._comboBox_type = combo_box_type
-        # Phase selected or not
-        self._checkBox_selected = check_box_select
-
-        # set up the unit cell dictionary
-        self._cellTypeList = list()
-        for type_tup in UnitCellList:
-            self._cellTypeList.append(type_tup[0])
-
-        combo_box_type.currentIndexChanged.connect(self.event_space_group_changed)
-        # parent.connect(combo_box_type, QtCore.SIGNAL('currentIndexChanged(int)'),
-        #                self.event_space_group_changed)
-
-        return
-
-    def enable_widgets(self, enabled):
-        """
-
-        :param enabled:
-        :return:
-        """
-        assert isinstance(enabled, bool)
-
-        self._lineEdit_a.setEnabled(enabled)
-        self._lineEdit_b.setEnabled(enabled)
-        self._lineEdit_c.setEnabled(enabled)
-        self._comboBox_type.setEnabled(enabled)
-        self._lineEdit_name.setEnabled(enabled)
-
-        return
-
-    def event_space_group_changed(self):
-        """
-        Purpose: handle the change of space group
-        Requirements:
-        :return:
-        """
-        curr_space_group = str(self._comboBox_type.currentText())
-
-        cell_type = curr_space_group.split()[0]
-
-        self.set_unit_cell_type(cell_type)
-
-        return
-
-    def get_phase_value(self, phase_value_list):
-        """
-        Purpose: set the phase values to the input list. It is used for save_to_buffer the values of phase temporarily.
-            if value is not set up, then it will ignored;
-        Requirements:  if the value is set, then it must be valid
-
-        :param phase_value_list:
-        :return:
-        """
-        # Check requirements:
-        assert isinstance(phase_value_list, list), 'Phase value list %s must be a list but not of type %s.' \
-                                                   '' % (str(phase_value_list), type(phase_value_list))
-        assert len(phase_value_list) == 5, 'Phase value list %s must be 5 elements.' % str(phase_value_list)
-
-        # Set name
-        phase_value_list[0] = str(self._lineEdit_name.text()).strip()
-        # Set phase
-        phase_value_list[1] = str(self._comboBox_type.currentText()).split()[0]
-        # Set a, b and c
-        a = GuiUtility.parse_float(self._lineEdit_a)
-        phase_value_list[2] = a
-        b = GuiUtility.parse_float(self._lineEdit_b)
-        phase_value_list[3] = b
-        c = GuiUtility.parse_float(self._lineEdit_c)
-        phase_value_list[4] = c
-
-        return
-
-    def get_phase(self):
-        """
-        Get the phase's parameters
-        Requirements:
-        1. a, b and c are positive floats
-        2. phase name must be given
-        :return: list as [name, type, a, b, c]
-        """
-        name = str(self._lineEdit_name.text()).strip()
-        assert len(name) > 0, 'Phase name must be given!'
-
-        cell_type = str(self._comboBox_type.currentText()).split()[0]
-        try:
-            a = float(self._lineEdit_a.text())
-            if self._lineEdit_b.isEnabled():
-                b = float(self._lineEdit_b.text())
-            else:
-                b = a
-            if self._lineEdit_c.isEnabled():
-                c = float(self._lineEdit_c.text())
-            else:
-                c = a
-        except TypeError as e:
-            raise RuntimeError('Lattice parameters a, b or c does not have correct value. Error: %s.' % str(e))
-
-        return [name, cell_type, a, b, c]
-
-    def is_selected(self):
-        """
-        Return the flag whether this phase is selected
-        :return:
-        """
-        return self._checkBox_selected.isChecked()
-
-    def reset(self):
-        """
-        Reset the widgets
-        :return:
-        """
-        # Clear name, a, b, c
-        self._lineEdit_a.setText('')
-        self._lineEdit_b.setText('')
-        self._lineEdit_c.clear()
-        self._lineEdit_name.setText('')
-
-        # Set the unit type to primitive
-        self.set_unit_cell_type('Primitive')
-
-        return
-
-    def set_lattice_widgets_values(self, cell_type):
-        """
-        Purpose:
-            enable or disabled some widgets according to unit cell type
-            set the values to disabled lattice parameters
-        :param cell_type:
-        :return:
-        """
-        assert cell_type in self._cellTypeList, 'Unit cell type %s is not supported.' % cell_type
-
-        if cell_type in ['BCC', 'FCC']:
-            # Disable inputs for b and c
-            self._lineEdit_b.setEnabled(False)
-            self._lineEdit_c.setEnabled(False)
-            self._lineEdit_b.setText('')
-            self._lineEdit_c.setText('')
-        elif cell_type in ['HCP']:
-            # Disable b and enable c
-            self._lineEdit_b.setEnabled(False)
-            self._lineEdit_c.setEnabled(True)
-            self._lineEdit_b.setText('')
-        else:
-            # enable all
-            self._lineEdit_b.setEnabled(True)
-            self._lineEdit_c.setEnabled(True)
-
-        return
-
-    def set_unit_cell_type(self, cell_type):
-        """
-        Set unit cell type and enable/disable the line edits for a, b and c
-        :param cell_type:
-        :return:
-        """
-        # Check
-        assert isinstance(cell_type, str)
-        assert cell_type in self._cellTypeList, 'Unit cell type %s is not supported.' % cell_type
-
-        # Set
-        list_index = self._cellTypeList.index(cell_type)
-        self._comboBox_type.setCurrentIndex(list_index)
-
-        # Disable some
-        self.set_lattice_widgets_values(cell_type)
-
-        return
-
-    def set_value(self, unit_cell_value):
-        """ Set value to unit cell (usually for undo)
-        Purpose: set the values of unit cell (stored in list) to widgets
-        Requirements: unit cell value should be stored in a list with 5 elements
-        Guarantees: unit cell values are set up
-        :param unit_cell_value: 5-element list
-        :return:
-        """
-        # Check requirements
-        assert isinstance(unit_cell_value, list), 'Unit cell value %s must be given in list but not %s.' \
-                                                  '' % (str(unit_cell_value), type(unit_cell_value))
-        assert len(unit_cell_value) == 5, 'Unit cell value %s must have 5 elements.' % str(unit_cell_value)
-
-        # Set phase name
-        self._lineEdit_name.setText(unit_cell_value[0])
-        # Set a, b and c
-        if unit_cell_value[2] is None:
-            self._lineEdit_a.clear()
-        else:
-            self._lineEdit_a.setText(str(unit_cell_value[2]))
-        if unit_cell_value[3] is None:
-            self._lineEdit_a.clear()
-        else:
-            self._lineEdit_a.setText(str(unit_cell_value[3]))
-        if unit_cell_value[4] is None:
-            self._lineEdit_a.clear()
-        else:
-            self._lineEdit_a.setText(str(unit_cell_value[4]))
-        # Set unit cell type
-        new_index = -1
-        for index in xrange(len(UnitCellList)):
-            if unit_cell_value[1] == UnitCellList[index][0]:
-                new_index = index
-                break
-        if new_index == -1:
-            raise RuntimeError('Impossible to find unit cell type %s not in the list.' % unit_cell_value[1])
-        else:
-            self._comboBox_type.setCurrentIndex(new_index)
-
-        return
 
 
 class PeakPickerMode(object):
@@ -413,9 +64,7 @@ class PeakPickerWindow(QMainWindow):
         # Define event handling methods
         # phase set up
         self.ui.pushButton_setPhases.clicked.connect(self.do_set_phases)
-
         self.ui.pushButton_clearPhase.clicked.connect(self.do_clear_phases)
-
         self.ui.pushButton_cancelPhaseChange.clicked.connect(self.do_undo_phase_changes)
 
         # peak processing
@@ -444,8 +93,8 @@ class PeakPickerWindow(QMainWindow):
 
         self.ui.pushButton_peakPickerMode.clicked.connect(self.do_set_pick_mode)
 
-        # TODO - NIGHT - Implement
-        # self.ui.pushButton_clearGroup
+        self.ui.pushButton_clearGroup.clicked.connect(self.do_clear_groups)
+        self.ui.pushButton_resetSelection.clicked.connect(self.do_clear_peak_selection)
 
         # vanadium
         self.ui.pushButton_launchVanProcessDialog.clicked.connect(self.do_launch_vanadium_dialog)
@@ -463,9 +112,6 @@ class PeakPickerWindow(QMainWindow):
 
         # get terminal
         self.ui.actionLaunch_Terminal.triggered.connect(self.menu_launch_terminal)
-
-        # TODO - NIGHT - Implement features
-        # pushButton_resetSelection : reset selected peaks
 
         # self.connect(self.ui.pushButton_setPhases, QtCore.SIGNAL('clicked()'),
         #              self.do_set_phases)
@@ -687,11 +333,11 @@ class PeakPickerWindow(QMainWindow):
 
     def add_grouped_peaks(self):
         """
-
         :return:
         """
         # check
-        assert self._autoPeakGroup is not None
+        if self._autoPeakGroup is None:
+            raise RuntimeError('_autoPeakGroup shall be set up before this method is called')
 
         # get information
         # get bank
@@ -699,21 +345,22 @@ class PeakPickerWindow(QMainWindow):
 
         # get number of groups
         group_id_list = sorted(self._autoPeakGroup.get_group_ids())
-        num_groups = len(group_id_list)
+        # num_groups = len(group_id_list)
+
+        # clear table as there is NO PEAK ADDITION mode to the table and thus the nameing
+        self.ui.tableWidget_peakParameter.clear_selected_peaks()
 
         # add peak to table
-        for group_id in group_id_list:
-
+        for index, group_id in enumerate(group_id_list):
             peak_name = ''
-
             group_left_b, group_right_b = self._autoPeakGroup.get_fit_range(group_id)
             width = group_right_b - group_left_b
 
             peak_tup_list = self._autoPeakGroup.get_peaks(group_id)
+            print ('[DB...BAT] {}-group with ID {}: {}'.format(index, group_id, peak_tup_list))
 
             for peak_tup in peak_tup_list:
                 peak_centre = peak_tup[1]
-
                 self.ui.tableWidget_peakParameter.add_peak(bank=bank, name=peak_name,
                                                            centre=peak_centre,
                                                            width=width,
@@ -871,6 +518,31 @@ class PeakPickerWindow(QMainWindow):
 
         return
 
+    def do_clear_groups(self):
+        """ Clean grouped peaks
+        :return:
+        """
+        # clear table
+        self.ui.tableWidget_peakParameter.clear_selected_peaks()
+
+        # clear the indicators on the image
+        self.ui.graphicsView_main.clear_highlight_data()
+
+        return
+
+    def do_clear_peak_selection(self):
+        """
+        clear picked peaks
+        :return:
+        """
+        print ('[DB...BAT] Clear groups first')
+        self.do_clear_groups()
+
+        # remove all from Diffraction Plot View
+        self.ui.graphicsView_main.reset_selected_peaks()
+
+        return
+
     def do_clear_peaks_fm_canvas(self):
         """
         Purpose:
@@ -888,9 +560,7 @@ class PeakPickerWindow(QMainWindow):
         raise NotImplementedError('Add button to GUI')
 
     def do_delete_peaks(self):
-        """
-        Purpose:
-            Delete the selected peak from table and place holder and their indicators
+        """ Delete the selected peak from table and place holder and their indicators
         Requirements:
             At least one peak is selected in the table
         Guarantees:
@@ -899,7 +569,9 @@ class PeakPickerWindow(QMainWindow):
         """
         # Get the rows that contain the peaks to delete
         row_number_list = self.ui.tableWidget_peakParameter.get_selected_rows()
-        assert len(row_number_list) > 0, 'No peak is selected to delete.'
+        if len(row_number_list) > 0:
+            GuiUtility.pop_dialog_information(self, 'No peak is selected to delete')
+            return
 
         # Delete the selected rows
         self.ui.tableWidget_peakParameter.remove_rows(row_number_list)
@@ -1073,7 +745,8 @@ class PeakPickerWindow(QMainWindow):
         by its position
         :return:
         """
-        self._groupPeakDialog = GroupPeakDialog.GroupPeakDialog(self)
+        if self._groupPeakDialog is None:
+            self._groupPeakDialog = GroupPeakDialog(self)
         self._groupPeakDialog.show()
 
         return
@@ -1361,7 +1034,7 @@ class PeakPickerWindow(QMainWindow):
             # Load data via parent
             try:
                 data_key = os.path.basename(gsas_file_name).split('_')[0] + 'H'
-                data_key = self._myController.load_diffraction_file(gsas_file_name, 'gsas', data_key)
+                data_key = self._myController.load_diffraction_file(gsas_file_name, 'gsas', data_key, unit='dSpacing')
                 self._dataKeyList.append(data_key)
                 # add to tree
                 # self.ui.treeView_iptsRun.add_child_current_item(data_key)
@@ -1584,8 +1257,12 @@ class PeakPickerWindow(QMainWindow):
         peak_width_dialog = PeakWidthSetupDialog(self)
         peak_width_dialog.exec_()
 
-        peak_width = peak_width_dialog.get_peak_width()
-        print 'Get log value = ', peak_width
+        try:
+            peak_width = peak_width_dialog.get_peak_width()
+            # TODO - WHAT TO DO WITH SET PEAKS WITH?
+            print ('Peaks width is set to {}'.format(peak_width))
+        except AssertionError:
+            pass
 
         return
 
@@ -1649,6 +1326,7 @@ class PeakPickerWindow(QMainWindow):
 
         return
 
+    # TODO - NIGHT - Clean
     def group_peaks(self, resolution, num_fwhm):
         """
         Group a list of peaks for fitting
@@ -1667,6 +1345,7 @@ class PeakPickerWindow(QMainWindow):
         assert isinstance(peak_group, peak_util.PeakGroupCollection),\
             'Peak group {0} must be a PeakGroupCollection instance but not a {1}.'.format(peak_group, type(peak_group))
 
+        # clear previous grouped peaks' presentation on PLOT
         self.clear_group_highlight()
 
         # reflect the grouped peak to GUI
@@ -2108,29 +1787,3 @@ class PeakPickerWindow(QMainWindow):
         self._myController.undo_vanadium_smoothing()
 
         return
-
-
-
-def retrieve_peak_positions(peak_tup_list):
-    """
-    Purpose:
-        Retrieve peak positions from peaks in given list
-    Requirements:
-        Input is a list of 2-tuples as HKL and d-spacing
-    Guarantees:
-        Retrieve the peaks' positions out
-    :param peak_tup_list:
-    :return: a list of
-    """
-    assert isinstance(peak_tup_list, list), 'Peak tuple list should be a list but not of type ' \
-                                            '%s.' % str(type(list))
-
-    peak_pos_list = list()
-    for peak in peak_tup_list:
-        peak_info_tup = peak[0]
-        peak_pos = peak_info_tup[0]
-        peak_pos_list.append(peak_pos)
-    # END-FOR(peak)
-
-    return peak_pos_list
-
